pyproject.toml
[project]
name = "agendint-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agendint",
    "apscheduler>=3.11.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
]

[tool.uv.sources]
agendint = { git = "https://github.com/INT-Scripts/int-libraries.git", subdirectory = "packages/agendint" }
